pyproject.toml
[project]
name = "psof"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.8",
    "mypy>=1.15.0",
    "psycopg2-binary>=2.9.10",
    "pydantic-settings>=2.8.1",
    "sqlalchemy>=2.0.40",
    "uvicorn>=0.34.0",
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"

[dependency-groups]
dev = [
    "pytest-asyncio>=0.25.3",
    "pytest>=8.3.5",
    "pytest-cov>=6.0.0",
    "httpx>=0.28.1",
    "alembic>=1.15.2",
]
